refactor: split runtime out of main.rs and fix native playback metadata (#218)
