+++ src/core/mod.rs
@@ -1,4 +1,5 @@
 pub mod app;
+pub mod auth;
 pub mod config;
 pub mod layout;
 pub mod sort;
