+++ src/infra/player/mod.rs
@@ -4,7 +4,11 @@
 //! It registers spotatui as a Spotify Connect device and handles audio streaming.
 
 #[cfg(feature = "streaming")]
+mod events;
+#[cfg(feature = "streaming")]
 mod streaming;
 
+#[cfg(feature = "streaming")]
+pub use events::*;
 #[cfg(feature = "streaming")]
 pub use streaming::*;
