+++ src/tui/handlers/error_screen.rs
@@ -1,3 +1,3 @@
-use crate::{app::App, event::Key};
+use crate::{core::app::App, tui::event::Key};
 
 pub fn handler(_key: Key, _app: &mut App) {}
