+++ src/tui/mod.rs
@@ -3,4 +3,5 @@ pub mod banner;
 pub mod cover_art;
 pub mod event;
 pub mod handlers;
+pub mod runner;
 pub mod ui;
